Added DataCenterCViewDelegate to allow for different handling of keystrokes and mice input.

+++ libcview/DataCenter/Node.h
@@ -15,8 +15,8 @@
     BOOL fading;
     BOOL unfading;
     BOOL wasfading;
-    float fadetime;
-    float fadestart;
+    double fadetime;
+    double fadestart;
     double fadeval;
 }
 +(void)setNodeArray:(VertArray*)_nodeArray;
